+++ lib/controller/controller_manager.dart
@@ -10,7 +10,6 @@ import 'package:chat_interface/controller/conversation/spaces/spaces_controller.
 import 'package:chat_interface/controller/conversation/conversation_controller.dart';
 import 'package:chat_interface/controller/conversation/message_controller.dart';
 import 'package:chat_interface/controller/conversation/spaces/spaces_member_controller.dart';
-import 'package:chat_interface/controller/conversation/townsquare_controller.dart';
 import 'package:chat_interface/controller/current/connection_controller.dart';
 import 'package:chat_interface/controller/current/status_controller.dart';
 import 'package:chat_interface/database/trusted_links.dart';
@@ -39,7 +38,6 @@ void initializeControllers() {
   Get.put(SettingController());
   Get.put(ThemeManager());
   Get.put(TransitionController());
-  Get.put(TownsquareController());
 
   // Space controls
   Get.put(SpacesController());