+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Added an account management page that talks to the new endpoints for search, deleting and changing the ranks of accounts
 - Changed the app icon to better work with white mode on Windows (color still isn't good, but it looks a little better)
 - New settings for admins in the Town page in the settings (fully server-rendered, expect different settings to be added over time)
+- Fixed a bug where files wouldn't be downloaded because they are below the auto-download size (when pressing the download button)